Updated code to use Gameplay Effects
Updated AuraEffectActor to take advantage of GameplayEffects, the class now supports Instant, Duration and Infinite based effects. Also added the support for ActorLevels enabling the use of curves.

Added assets for using the different types of gameplay effects such as a FireArea which will deplete heal as long as someone is standing on it using an infinite effect. Health/Mana Potions which have an instant effect and crystals which use a duration effect to increase either health or mana over a specified time.